+++ parsley/src/ghc/Parsley/Fold.hs
@@ -16,18 +16,18 @@ module Parsley.Fold (
     skipMany, skipSome, skipManyN, --loop,
     sepBy, sepBy1, endBy, endBy1, sepEndBy, sepEndBy1,
     chainl1, chainr1, chainl, chainr,
-    chainl1', chainr1', chainPre, chainPost,
-    pfoldr, pfoldl,
-    pfoldr1, pfoldl1
+    infixl1, infixr1, prefix, postfix,
+    manyr, manyl,
+    somer, somel
   ) where
 
 import Prelude hiding           (pure, (<*>), (<$>), (*>), (<*))
 import Parsley.Alternative      ((<|>), option)
 import Parsley.Applicative      (pure, (<*>), (<$>), (*>), (<*), (<:>), (<**>), void)
-import Parsley.Defunctionalized (Defunc(FLIP, ID, COMPOSE, EMPTY, CONS, CONST), pattern FLIP_H, pattern COMPOSE_H, pattern UNIT)
+import Parsley.Defunctionalized (Defunc(FLIP, ID, COMPOSE, EMPTY, CONS, CONST, APP_H), pattern FLIP_H, pattern UNIT)
 import Parsley.Internal         (Parser)
 import Parsley.ParserOps        (ParserOps)
-import Parsley.Register         (bind, get, put, modify, newRegister, newRegister_)
+import Parsley.Register         (get, modify, newRegister, newRegister_)
 
 import qualified Parsley.Internal as Internal (loop)
 
@@ -37,7 +37,7 @@ failed having not consumed input, @exit@ is performed, otherwise the combinator 
 
 > loop body exit = let go = body *> go <|> exit in go
 
-@since 1.1.0.0
+@since 2.0.0.0
 -}
 loop :: Parser () -> Parser a -> Parser a
 loop = Internal.loop
@@ -46,12 +46,12 @@ loop = Internal.loop
 This combinator parses repeated applications of an operator to a single final operand. This is
 primarily used to parse prefix operators in expressions.
 
-@since 0.1.0.0
+@since 2.0.0.0
 -}
-chainPre :: Parser (a -> a) -> Parser a -> Parser a
-chainPre op p =
-  newRegister (pure ID) (\r ->
-    loop (put r (pure (FLIP_H COMPOSE) <*> op <*> get r))
+prefix :: Parser (a -> a) -> Parser a -> Parser a
+prefix op p =
+  newRegister_ ID (\r ->
+    loop (modify r (FLIP_H COMPOSE <$> op))
          (get r))
   <*> p
 
@@ -59,99 +59,95 @@ chainPre op p =
 This combinator parses repeated applications of an operator to a single initial operand. This is
 primarily used to parse postfix operators in expressions.
 
-@since 0.1.0.0
+@since 2.0.0.0
 -}
-chainPost :: Parser a -> Parser (a -> a) -> Parser a
-chainPost p op =
+postfix :: Parser a -> Parser (a -> a) -> Parser a
+postfix p op =
   newRegister p $ \r ->
-    loop (put r (op <*> get r))
+    loop (modify r op)
          (get r)
 
 -- Parser Folds
 {-|
-@pfoldr f k p@ parses __zero__ or more @p@s and combines the results using the function @f@. When @p@
+@manyr f k p@ parses __zero__ or more @p@s and combines the results using the function @f@. When @p@
 fails without consuming input, the terminal result @k@ is returned.
 
-> many = pfoldr CONS EMPTY
+> many = manyr CONS EMPTY
 
-@since 0.1.0.0
+@since 2.0.0.0
 -}
-pfoldr :: (ParserOps repf, ParserOps repk) => repf (a -> b -> b) -> repk b -> Parser a -> Parser b
-pfoldr f k p = chainPre (f <$> p) (pure k)
+manyr :: (ParserOps repf, ParserOps repk) => repf (a -> b -> b) -> repk b -> Parser a -> Parser b
+manyr f k p = prefix (f <$> p) (pure k)
 
 {-|
-@pfoldr1 f k p@ parses __one__ or more @p@s and combines the results using the function @f@. When @p@
+@somer f k p@ parses __one__ or more @p@s and combines the results using the function @f@. When @p@
 fails without consuming input, the terminal result @k@ is returned.
 
-> some = pfoldr1 CONS EMPTY
+> some = somer CONS EMPTY
 
-@since 0.1.0.0
+@since 2.0.0.0
 -}
-pfoldr1 :: (ParserOps repf, ParserOps repk) => repf (a -> b -> b) -> repk b -> Parser a -> Parser b
-pfoldr1 f k p = f <$> p <*> pfoldr f k p
+somer :: (ParserOps repf, ParserOps repk) => repf (a -> b -> b) -> repk b -> Parser a -> Parser b
+somer f k p = f <$> p <*> manyr f k p
 
 {-|
-@pfoldl f k p@ parses __zero__ or more @p@s and combines the results using the function @f@. The
+@manyl f k p@ parses __zero__ or more @p@s and combines the results using the function @f@. The
 accumulator is initialised with the value @k@.
 
-@since 0.1.0.0
+@since 2.0.0.0
 -}
-pfoldl :: (ParserOps repf, ParserOps repk) => repf (b -> a -> b) -> repk b -> Parser a -> Parser b
-pfoldl f k p = chainPost (pure k) ((FLIP <$> pure f) <*> p)
+manyl :: (ParserOps repf, ParserOps repk) => repf (b -> a -> b) -> repk b -> Parser a -> Parser b
+manyl f k p = postfix (pure k) ((FLIP <$> pure f) <*> p)
 
 {-|
-@pfoldl1 f k p@ parses __one__ or more @p@s and combines the results using the function @f@. The
+@somel f k p@ parses __one__ or more @p@s and combines the results using the function @f@. The
 accumulator is initialised with the value @k@.
 
-@since 0.1.0.0
+@since 2.0.0.0
 -}
-pfoldl1 :: (ParserOps repf, ParserOps repk) => repf (b -> a -> b) -> repk b -> Parser a -> Parser b
-pfoldl1 f k p = chainPost (f <$> pure k <*> p) ((FLIP <$> pure f) <*> p)
+somel :: (ParserOps repf, ParserOps repk) => repf (b -> a -> b) -> repk b -> Parser a -> Parser b
+somel f k p = postfix (f <$> pure k <*> p) ((FLIP <$> pure f) <*> p)
 
 -- Chain Combinators
 {-|
-@chainl1' wrap p op @ parses one or more occurrences of @p@, separated by @op@. Returns a value obtained
+@infixl1 wrap p op @ parses one or more occurrences of @p@, separated by @op@. Returns a value obtained
 by a /left/ associative application of all functions returned by @op@ to the values returned by @p@.
 The function @wrap@ is used to transform the initial value from @p@ into the correct form.
 
-@since 0.1.0.0
+@since 2.0.0.0
 -}
-chainl1' :: ParserOps rep => rep (a -> b) -> Parser a -> Parser (b -> a -> b) -> Parser b
-chainl1' f p op = chainPost (f <$> p) (FLIP <$> op <*> p)
+infixl1 :: ParserOps rep => rep (a -> b) -> Parser a -> Parser (b -> a -> b) -> Parser b
+infixl1 wrap p op = postfix (wrap <$> p) (FLIP <$> op <*> p)
 
 {-|
-The classic version of the left-associative chain combinator. See 'chainl1''.
+The classic version of the left-associative chain combinator. See 'infixl1'.
 
-> chainl1 p op = chainl1' ID p op
+> chainl1 p op = infixl1 ID p op
 
 @since 0.1.0.0
 -}
 chainl1 :: Parser a -> Parser (a -> a -> a) -> Parser a
-chainl1 = chainl1' ID
+chainl1 = infixl1 ID
 
 {-|
-@chainr1' wrap p op @ parses one or more occurrences of @p@, separated by @op@. Returns a value obtained
+@infixr1 wrap p op @ parses one or more occurrences of @p@, separated by @op@. Returns a value obtained
 by a /right/ associative application of all functions returned by @op@ to the values returned by @p@.
 The function @wrap@ is used to transform the final value from @p@ into the correct form.
 
-@since 0.1.0.0
+@since 2.0.0.0
 -}
-chainr1' :: ParserOps rep => rep (a -> b) -> Parser a -> Parser (a -> b -> b) -> Parser b
-chainr1' f p op = newRegister_ ID $ \acc ->
-  let go = bind p $ \x ->
-           modify acc (FLIP_H COMPOSE <$> (op <*> x)) *> go
-       <|> f <$> x
-  in go <**> get acc
+infixr1 :: ParserOps rep => rep (a -> b) -> Parser a -> Parser (a -> b -> b) -> Parser b
+infixr1 wrap p op = let go = p <**> (FLIP <$> op <*> go <|> pure wrap) in go
 
 {-|
-The classic version of the right-associative chain combinator. See 'chainr1''.
+The classic version of the right-associative chain combinator. See 'infixr1'.
 
-> chainr1 p op = chainr1' ID p op
+> chainr1 p op = infixr1 ID p op
 
 @since 0.1.0.0
 -}
 chainr1 :: Parser a -> Parser (a -> a -> a) -> Parser a
-chainr1 = chainr1' ID
+chainr1 = infixr1 ID
 
 {-|
 Like `chainr1`, but may parse zero occurences of @p@ in which case the value is returned.
@@ -177,7 +173,7 @@ into a list. Same as @manyN 0@
 @since 0.1.0.0
 -}
 many :: Parser a -> Parser [a]
-many = pfoldr CONS EMPTY
+many = manyr CONS EMPTY
 
 {-|
 Attempts to parse the given parser __n__ or more times, collecting all of the successful results
@@ -203,9 +199,8 @@ Like `many`, excepts discards its results.
 @since 0.1.0.0
 -}
 skipMany :: Parser a -> Parser ()
---skipMany p = let skipManyp = p *> skipManyp <|> unit in skipManyp
-skipMany = void . pfoldl CONST UNIT -- the void here will encourage the optimiser to recognise that the register is unused
---skipMany = flip loop (pure UNIT) . void -- This should be the optimal one, with register removed, but apparently not?! something is amiss, perhaps space leak?
+--skipMany p = loop (void p) unit
+skipMany = void . manyl CONST UNIT -- This is still faster, the above generates better code, but GHC starts doing weird things!
 
 {-|
 Like `manyN`, excepts discards its results.
@@ -275,13 +270,7 @@ Returns a list of values returned by @p@.
 @since 0.1.0.0
 -}
 sepEndBy1 :: Parser a -> Parser b -> Parser [a]
-sepEndBy1 p sep = newRegister_ ID $ \acc ->
-  let go = modify acc (COMPOSE_H (FLIP_H COMPOSE) CONS <$> p)
-         *> (sep *> (go <|> get acc) <|> get acc)
-  in go <*> pure EMPTY
-
-{-sepEndBy1 :: Parser a -> Parser b -> Parser [a]
 sepEndBy1 p sep =
   let seb1 = p <**> (sep *> (FLIP_H CONS <$> option EMPTY seb1)
                  <|> pure (APP_H (FLIP_H CONS) EMPTY))
-  in seb1-}
+  in seb1